backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-filter>=23.2
django-cors-headers>=4.1.0
Pillow>=9.5.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
openai>=1.0.0
tiktoken
python-docx
pdfplumber
pandas
ebooklib
beautifulsoup4
markdown
openpyxl 
numpy>=1.24.0
sentence-transformers>=2.2.2